Убрали ноль после целого значения в результате

+++ src/CreditPaymentService.java
@@ -1,11 +1,11 @@
 public class CreditPaymentService {
-    public float calculate(float summa, float period) {
+    public int calculate(float summa, float period) {
         float result;
         float monthlyPercentage;
         {
             monthlyPercentage = (float) 9.99 / (100 * 12);
             result = (int) (summa * (monthlyPercentage / (1 - Math.pow(1 + monthlyPercentage, -period))));
         }
-        return result;
-    }
+        return (int) result;
+            }
 }
